Refactor the nomination code. Add more tests
